@@ -1,6 +1,7 @@
-import React, { useRef } from 'react';
+import React, { useRef, useState } from 'react';
 import Slider from 'react-slick';
-import SimpleReactLightbox, { SRLWrapper } from 'simple-react-lightbox';
+import Lightbox from 'yet-another-react-lightbox';
+import 'yet-another-react-lightbox/styles.css';
 import 'slick-carousel/slick/slick-theme.css';
 import 'slick-carousel/slick/slick.css';
 import shape12 from '../../assets/images/shape/shape-12.png';
@@ -15,6 +16,21 @@ import showCaseFive from '../../assets/images/showcase-5.png';
 
 function ShowCaseHomeThree() {
     const sliderRef = useRef();
+    const [lightboxOpen, setLightboxOpen] = useState(false);
+    const [currentIndex, setCurrentIndex] = useState(0);
+
+    const images = [
+        { src: showCaseOne },
+        { src: showCaseTwo },
+        { src: showCaseThree },
+        { src: showCaseFour },
+        { src: showCaseFive },
+    ];
+
+    const openLightbox = (index) => {
+        setCurrentIndex(index);
+        setLightboxOpen(true);
+    };
     // const sliderNext = () => {
     //     sliderRef.current.slickNext();
     // };
@@ -44,74 +60,55 @@ function ShowCaseHomeThree() {
     return (
         <>
             <section className="appie-showcase-area">
-                <SimpleReactLightbox>
-                    <SRLWrapper>
-                        <div className="container">
-                            <div className="row">
-                                <div className="col-lg-12">
-                                    <div className="appie-section-title text-center">
-                                        <h3 className="appie-title">Creative app showcase</h3>
-                                        <p>The app provides design and digital marketing.</p>
+                <div className="container">
+                    <div className="row">
+                        <div className="col-lg-12">
+                            <div className="appie-section-title text-center">
+                                <h3 className="appie-title">Creative app showcase</h3>
+                                <p>The app provides design and digital marketing.</p>
+                            </div>
+                        </div>
+                    </div>
+                    <div className="row appie-showcase-slider">
+                        <div className="col-lg-12">
+                            <Slider ref={sliderRef} {...settings}>
+                                {images.map((image, index) => (
+                                    <div key={index}>
+                                        <div className="appie-showcase-item mt-30">
+                                            <a 
+                                                className="appie-image-popup"
+                                                onClick={() => openLightbox(index)}
+                                                style={{ cursor: 'pointer' }}
+                                            >
+                                                <img src={image.src} alt="" />
+                                            </a>
+                                        </div>
                                     </div>
-                                </div>
-                            </div>
-                            <div className="row appie-showcase-slider">
-                                <div className="col-lg-12">
-                                    <Slider ref={sliderRef} {...settings}>
-                                        <div>
-                                            <div className="appie-showcase-item mt-30">
-                                                <a className="appie-image-popup">
-                                                    <img src={showCaseOne} alt="" />
-                                                </a>
-                                            </div>
-                                        </div>
-                                        <div>
-                                            <div className="appie-showcase-item mt-30">
-                                                <a className="appie-image-popup">
-                                                    <img src={showCaseTwo} alt="" />
-                                                </a>
-                                            </div>
-                                        </div>
-                                        <div>
-                                            <div className="appie-showcase-item mt-30">
-                                                <a className="appie-image-popup">
-                                                    <img src={showCaseThree} alt="" />
-                                                </a>
-                                            </div>
-                                        </div>
-                                        <div>
-                                            <div className="appie-showcase-item mt-30">
-                                                <a className="appie-image-popup">
-                                                    <img src={showCaseFour} alt="" />
-                                                </a>
-                                            </div>
-                                        </div>
-                                        <div>
-                                            <div className="appie-showcase-item mt-30">
-                                                <a className="appie-image-popup">
-                                                    <img src={showCaseFive} alt="" />
-                                                </a>
-                                            </div>
-                                        </div>
-                                    </Slider>
-                                </div>
-                            </div>
+                                ))}
+                            </Slider>
                         </div>
-                        <div className="showcase-shape-1">
-                            <img src={shape14} alt="" />
-                        </div>
-                        <div className="showcase-shape-2">
-                            <img src={shape13} alt="" />
-                        </div>
-                        <div className="showcase-shape-3">
-                            <img src={shape12} alt="" />
-                        </div>
-                        <div className="showcase-shape-4">
-                            <img src={shape15} alt="" />
-                        </div>
-                    </SRLWrapper>
-                </SimpleReactLightbox>
+                    </div>
+                </div>
+                <div className="showcase-shape-1">
+                    <img src={shape14} alt="" />
+                </div>
+                <div className="showcase-shape-2">
+                    <img src={shape13} alt="" />
+                </div>
+                <div className="showcase-shape-3">
+                    <img src={shape12} alt="" />
+                </div>
+                <div className="showcase-shape-4">
+                    <img src={shape15} alt="" />
+                </div>
             </section>
+            
+            <Lightbox
+                open={lightboxOpen}
+                close={() => setLightboxOpen(false)}
+                slides={images}
+                index={currentIndex}
+            />
         </>
     );
 }
